requirements.txt
httpx>=0.27.0
pyyaml>=6.0       # optional: manifest_loader falls back to minimal parser without it
